refactor: 魚の生成と移動ロジックを分離
